Feat : 주말머지정리 230911 0936

+++ Front-End/pro/app/src/main/java/com/example/project/viewmodels/TextloginViewModel.kt
@@ -29,7 +29,7 @@ class TextloginViewModel @Inject constructor(
                     response.body()?.let {
                         // SharedPreferences에 로그인 정보 저장
                         sharedPreferencesUtil.setLoggedInStatus(true)
-                        sharedPreferencesUtil.setUserId(it.user_idx.toInt())
+                        sharedPreferencesUtil.setUserId(it.user_idx)
                         sharedPreferencesUtil.setUserNickname(it.user_nickname)
                         sharedPreferencesUtil.setUserToken(it.user_accesstoken)
